# anms-core/integration_test/openapi.json
{
  "openapi": "3.0.2",
  "info": {
      "title": "AMMOS-ANMS",
      "description": "AMMOS-ANMS Docs",
      "version": "0.1.0"
  },
  "paths": {
...
      "/users/{username}": {
          "get": {
              "tags": [
                  "USER"
              ],
              "summary": "Get User",
              "operationId": "get_user_users__username__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Username",
                          "type": "string"
                      },
                      "name": "username",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/UserReturn"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          },
          "put": {
              "tags": [
                  "USER"
              ],
              "summary": "Update User",
              "operationId": "update_user_users__username__put",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Username",
                          "type": "string"
                      },
                      "name": "username",
                      "in": "path"
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/UserUpdateBase"
                          }
                      }
                  },
                  "required": true
              },
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {}
                          }
                      }
                  },
                  "409": {
                      "description": "Conflict",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Conflict"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/users": {
          "post": {
              "tags": [
                  "USER"
              ],
              "summary": "Create User",
              "operationId": "create_user_users_post",
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/UserBase"
                          }
                      }
                  },
                  "required": true
              },
              "responses": {
                  "201": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/UserReturn"
                              }
                          }
                      }
                  },
                  "409": {
                      "description": "Conflict",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Conflict"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
  },
  "components": {
      "schemas": {
...
          "Conflict": {
              "title": "Conflict",
              "required": [
                  "field"
              ],
              "type": "object",
              "properties": {
                  "field": {
                      "title": "Field",
                      "type": "object",
                      "example": {
                          "email": "abcd@example.com"
                      }
                  }
              }
          },
...
          "HTTPValidationError": {
              "title": "HTTPValidationError",
              "type": "object",
              "properties": {
                  "detail": {
                      "title": "Detail",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/ValidationError"
                      }
                  }
              }
          },
...
          "UserBase": {
              "title": "UserBase",
              "type": "object",
              "properties": {
                  "first_name": {
                      "title": "First Name",
                      "type": "string",
                      "example": "John"
                  },
                  "last_name": {
                      "title": "Last Name",
                      "type": "string",
                      "example": "Doe"
                  },
                  "email": {
                      "title": "Email",
                      "type": "string",
                      "format": "email",
                      "example": "john.doe@example.com"
                  },
                  "username": {
                      "title": "Username",
                      "type": "string",
                      "example": "johndoe"
                  }
              }
          },
          "UserReturn": {
              "title": "UserReturn",
              "required": [
                  "id",
                  "created_at",
                  "updated_at"
              ],
              "type": "object",
              "properties": {
                  "first_name": {
                      "title": "First Name",
                      "type": "string",
                      "example": "John"
                  },
                  "last_name": {
                      "title": "Last Name",
                      "type": "string",
                      "example": "Doe"
                  },
                  "email": {
                      "title": "Email",
                      "type": "string",
                      "format": "email",
                      "example": "john.doe@example.com"
                  },
                  "username": {
                      "title": "Username",
                      "type": "string",
                      "example": "johndoe"
                  },
                  "id": {
                      "title": "Id",
                      "type": "integer",
                      "example": 1
                  },
                  "created_at": {
                      "title": "Created At",
                      "type": "string",
                      "example": "2022-11-18T15:29:37.226127"
                  },
                  "updated_at": {
                      "title": "Updated At",
                      "type": "string",
                      "example": "2022-11-18T15:29:37.226127"
                  }
              }
          },
          "UserUpdateBase": {
              "title": "UserUpdateBase",
              "type": "object",
              "properties": {
                  "first_name": {
                      "title": "First Name",
                      "type": "string",
                      "example": "John"
                  },
                  "last_name": {
                      "title": "Last Name",
                      "type": "string",
                      "example": "Doe"
                  },
                  "email": {
                      "title": "Email",
                      "type": "string",
                      "format": "email",
                      "example": "john.doe@example.com"
                  }
              }
          },
          "ValidationError": {
              "title": "ValidationError",
              "required": [
                  "loc",
                  "msg",
                  "type"
              ],
              "type": "object",
              "properties": {
                  "loc": {
                      "title": "Location",
                      "type": "array",
                      "items": {
                          "anyOf": [
                              {
                                  "type": "string"
                              },
                              {
                                  "type": "integer"
                              }
                          ]
                      }
                  },
                  "msg": {
                      "title": "Message",
                      "type": "string"
                  },
                  "type": {
                      "title": "Error Type",
                      "type": "string"
                  }
              }
          },
...
      }
  }
}